ARTS-650: moved test-supporting classes into their own package
